Add ability to change ZBD zone processing criteria
Introduced a function pointer, can_process_zone(), to be called
instead of checking for offline zones explicitly in the code.
The function that is assigned by default does the same thing,
but this change makes possible to override the existing behavior
and check for some other zone-related criteria. For instance, some
zoned devices may use zone types and conditions that are not defined
in blkzoned.h. In this case, another component, such as libzbc
ioengine, may override the default function to support this
exteneded functionality.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>