Add support for Zone Domains in libzbc ioengine
Coded checks to properly recognize Zone Domains devices
(which will appear as regular zon-zoned devices to the
system).

If the backend device is indeed ZD, then the new code will
make sure that not only OFFLINE zones will be skipped for I/O,
but also INACTIVE ones as well as zones of type GAP. READONLY
zones are not chosen for write.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>