Trim zone range of Zone Domains devices
Added code to trim inactive and gap zones from the top of
zone range of Zone Domains devices. Depending on the current
state of the device, it can be a lot of such zones at the top
and they can cause some noticeable overhead from ZBD code trying
to redirect i/o from these zones to ones that can be actually used
for i/o.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>