Make zbd.c compile without blkzoned.h
Added conditional compilation statements to allow zbd.c to
compile without errors even if the build system doesn't have
zoned block device support enabled in the kernel. If it is not
enabled, linux/blkzoned.h is not going to be present and the
include and all the definitions from that file should be skipped.
In this scenario, the user will have to use a specialized ioengine,
such as libzbc, for fio to properly recognize a zoned block defvice
and initialize the code in zbd.c to appropriately handle zoned I/O.

Some earlier changes in configure and Makefile (not a part of this
commit) make sure that zbd.c is compiled when either zoned block
support or libzc is present in the system. If both are available,
the user can select psync or libaio ioengine to utilize the
functionality provided by Linux (via ioctl and sysfs) or they can
select libzbc ioengine, which is able to handle a wider variety of
zoned devices and is able to choose ZBD-related options
automatically, based on the capabilities of the zoned DUT. In
particular, if libzbc ioengine is used, it will set zonemode=zbd
option automatically when it recognizes a device as zoned.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>