Don't require zbc_version() in libzbc to detect it
Even though trying to call a library function is more reliable
when it comes to detecting if the library is installed, in case
of libzbc it would exclude any recent versions from libzbc master
from detection. Therefore, I've changed configure to only check
for the presense of zbc.h to detect libzbc.

Also, fixed a few merge problems and assigned a new version to
fio in this branch.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>