Don't override zonetype in libzbc ioengine
Changed zoned device scan code in libzbc ioengine not to set zonetype
to zbd if a zoned device is detected. One reason for this is that
users may explicitly set zonemode=strided and in this case we want
to keep this setting. When setting the "strided" option, the user will
have to specify the zonesize as well because of some option validation
that happens before opening of any files.

This commit also includes a minor code cleanup and debugging
improvements.

Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>